+++ R/GSODR.R
@@ -40,11 +40,13 @@
 #' @seealso \code{\link{nearest_stations}} Find stations within a given radius
 #' expressed in kilometres for a given point using Latitude and Longitude
 #'
-#' @seealso \code{\link{get_station_list}} Fetch latest list of stations and
-#' corresponding metadata from the NCEI FTP server
+#' @seealso \code{\link{get_station_list}} Fetch the latest list of stations and
+#' corresponding metadata from the NCEI FTP server and update the package's
+#' local copy with these data
 #'
-#' @seealso \code{\link{SRTM_GSOD_elevation}} GSDOR provides additional
-#' elevation data derived from SRTM90m data
+#' @seealso \code{\link{isd_history}} GSDOR provides a cleaned, local copy of
+#' the "isd_history.csv" file to help make some processes faster with 
+#' elevation data derived from SRTM90m data also included in these data
 #'
 #' @seealso \code{\link{country_list}} GSODR provides a cleaned list of the
 #' countries which are represented in the GSOD data set