+++ R/get_station_list.R
@@ -1,22 +1,36 @@
 #' Download the Latest Station List From the NCEI Server
 #'
-#' This function downloads the latest station list from the NCEI FTP server.
-#' This list includes metadata for all stations including unique identifiers,
-#' country, state (if in US), latitude, longitude, elevation and when weather
-#' observations begin and end.  Stations with invalid latitude and longitude
-#' values will not be included.
+#' This function downloads the latest station list (isd-history.csv) from the
+#' NCEI FTP server and updates the data distributed with \code{GSODR} so that
+#' you have the latest list of stations available.  These data provide unique
+#' identifiers, country, state (if in US), latitude, longitude, elevation and
+#' when weather observations begin and end.  Stations with invalid latitude and
+#' longitude values will not be included.
+#' 
+#' There is no need to use this unless you know that a station exists in the
+#' GSODR data that is not available in the database distributed with 
+#' \code{\link{GSODR}} in the \code{\link{isd_history}}
+#' data.
 #'
 #' @examples
 #' \dontrun{
-#' GSOD_stations <- get_station_list()
+#' GSOD_station_list <- get_station_list()
 #' }
 #' @return \code{\link[data.table]{data.table}} object of station metadata.
 #' @author Adam H Sparks, \email{adamhsparks@gmail.com}
 #' @export
 #'
 get_station_list <- function() {
-  STNID <- NULL
-  stations <- readr::read_csv(
+  original_timeout <- options("timeout")[[1]]
+  options(timeout = 300)
+  on.exit(options(timeout = original_timeout))
+  
+  # load current local copy of isd_history
+  utils::data("isd_history", package = "GSODR")
+  old_isd_history <- data.table::setDT(isd_history)
+  
+  # fetch new isd_history from NCEI server
+  new_isd_history <- readr::read_csv(
     "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
     col_types = "ccccccddddd",
     col_names = c(
@@ -34,19 +48,49 @@ get_station_list <- function() {
     ),
     skip = 1
   )
-  stations[stations == -999.9] <- NA
-  stations[stations == -999] <- NA
-  stations <- stations[stations$LAT != 0 & stations$LON != 0, ]
-  stations <- stations[stations$LAT > -90 & stations$LAT < 90, ]
-  stations <- stations[stations$LON > -180 & stations$LON < 180, ]
-  stations$STNID <-
-    as.character(paste(stations$USAF, stations$WBAN, sep = "-"))
-  SRTM_GSOD_elevation <-
-    data.table::setkey(GSODR::SRTM_GSOD_elevation, STNID)
-  data.table::setDT(stations)
-  data.table::setkey(stations, STNID)
-  stations <- stations[SRTM_GSOD_elevation, on = "STNID"]
-  stations <- stations[!is.na(stations$LAT), ]
-  stations <- stations[!is.na(stations$LON), ]
-  return(stations)
+  new_isd_history[new_isd_history == -999.9] <- NA
+  new_isd_history[new_isd_history == -999] <- NA
+  new_isd_history <-
+    new_isd_history[new_isd_history$LAT != 0 &
+                      new_isd_history$LON != 0,]
+  new_isd_history <-
+    new_isd_history[new_isd_history$LAT > -90 &
+                      new_isd_history$LAT < 90,]
+  new_isd_history <-
+    new_isd_history[new_isd_history$LON > -180 &
+                      new_isd_history$LON < 180,]
+  new_isd_history$STNID <-
+    as.character(paste(new_isd_history$USAF, new_isd_history$WBAN, sep = "-"))
+  new_isd_history <- new_isd_history[!is.na(new_isd_history$LAT),]
+  new_isd_history <- new_isd_history[!is.na(new_isd_history$LON),]
+  
+  # left join the old and new data
+  isd_history <- dplyr::left_join(
+    new_isd_history,
+    old_isd_history,
+    by = c(
+      "USAF",
+      "WBAN",
+      "STN_NAME",
+      "CTRY",
+      "STATE",
+      "CALL",
+      "LAT",
+      "LON",
+      "ELEV_M",
+      "BEGIN",
+      "END",
+      "STNID"
+    )
+  )
+  
+  isd_history <- data.table::setDT(isd_history)
+
+  # overwrite the existing isd_history.rda file on disk
+  pkg <- system.file(package = "GSODR")
+  path <-
+    file.path(file.path(pkg, "data"), paste0("isd_history.rda"))
+  save(isd_history, file = path, compress = "bzip2")
+  
+  return(isd_history)
 }